BrainTumorDetection/requirements.txt
torch
torchvision
streamlit
Pillow
matplotlib
opencv-python
